# REST-LIS/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "REST-LIS",
    "description": "An ASP.NET Core Web API for ordering tests and reporting results through UniversaLIS",
    "contact": {
      "name": "Roy Harmon",
      "url": "https://www.linkedin.com/in/roy-r-harmon/"
    },
    "license": {
      "name": "MIT License",
      "url": "https://github.com/roy-harmon/UniversaLIS/blob/main/LICENSE"
    },
    "version": "v1"
  },
  "paths": {
    "/reports/patients": {
      "get": {
        "tags": [
          "REST-LIS"
        ],
        "operationId": "GetPatients",
        "parameters": [
          {
            "name": "practicePatientID",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "patientName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "DOB",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "universalTestID",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Patient"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/reports/patients/{id}": {
      "get": {
        "tags": [
          "REST-LIS"
        ],
        "operationId": "GetPatientByID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/requests/patients": {
      "post": {
        "tags": [
          "REST-LIS"
        ],
        "operationId": "PostPatientOrders",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "get": {
        "tags": [
          "REST-LIS"
        ],
        "operationId": "GetPatientRequests",
        "parameters": [
          {
            "name": "practicePatientID",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "patientName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "DOB",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "universalTestID",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientRequest"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/requests/patients/{id}": {
      "get": {
        "tags": [
          "REST-LIS"
        ],
        "operationId": "GetPatientOrdersByID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientRequest"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "REST-LIS"
        ],
        "operationId": "UpdatePatientOrderByID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientRequest"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "REST-LIS"
        ],
        "operationId": "DeletePatientRequestByID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Patient": {
        "type": "object",
        "properties": {
          "patientID": {
            "type": "integer",
            "description": "The unique identifier assigned to the order record in the UniversaLIS internal database.",
            "format": "int32",
            "readOnly": true
          },
          "practicePatientID": {
            "type": "string",
            "description": "The unique ID assigned and used by the practice to identify the patient and his/her results upon return of the results of testing.",
            "nullable": true
          },
          "labPatientID": {
            "type": "string",
            "description": "The unique processing number assigned to the patient by the laboratory.",
            "nullable": true
          },
          "patientID3": {
            "type": "string",
            "description": "Optionally used for additional, universal, or manufacturer-defined identifiers.",
            "nullable": true
          },
          "patientName": {
            "type": "string",
            "description": "The patient’s name shall be presented in the following format: last name, first name, middle name or initial, suffix, and title, and each of these components shall be separated by a component delimiter (^).",
            "nullable": true
          },
          "mmName": {
            "type": "string",
            "description": "The optional mother’s maiden name may be required to distinguish between patients with the same birthdate and last name when registry files are very large.",
            "nullable": true
          },
          "dob": {
            "type": "string",
            "description": "The patient's birthdate, presented in YYYYMMDD format.",
            "format": "yyyyMMdd",
            "nullable": true
          },
          "sex": {
            "type": "string",
            "description": "This field shall be represented by M, F, or U.",
            "nullable": true
          },
          "race": {
            "type": "string",
            "description": "Abbreviation or full text names of ethnic groups may be entered. Note that multiple answers are permissible, separated by a component delimiter (^).",
            "nullable": true
          },
          "address": {
            "type": "string",
            "description": "The street address of the patient’s mailing address.",
            "nullable": true
          },
          "reserved": {
            "type": "string",
            "description": "This field is reserved for future expansion.",
            "nullable": true
          },
          "telNo": {
            "type": "string",
            "description": "The patient’s telephone number.",
            "nullable": true
          },
          "attendingPhysicianID": {
            "type": "string",
            "description": "The physician(s) caring for the patient as either names or codes, as agreed upon between the sender and the receiver. Identifiers or names, or both, should be separated by component delimiters (^) and formatted like the PatientName field. Multiple physician names (for example, ordering physician, attending physician, referring physician) shall be separated by repeat delimiters (\\).",
            "nullable": true
          },
          "special1": {
            "type": "string",
            "description": "Optional text field for vendor use (each laboratory can use this differently).",
            "nullable": true
          },
          "special2": {
            "type": "string",
            "description": "Optional text field for vendor use (each laboratory can use this differently).",
            "nullable": true
          },
          "height": {
            "type": "string",
            "description": "Optional numeric field containing the patient’s height. The default units are centimeters. If measured in terms of another unit, the units should also be transmitted separated by a component delimiter (^).",
            "nullable": true
          },
          "weight": {
            "type": "string",
            "description": "Optional numeric field containing the patient’s weight. The default units are kilograms. If measured in terms of another unit, for example, pounds, the unit name shall also be transmitted separated by a component delimiter (^).",
            "nullable": true
          },
          "diagnosis": {
            "type": "string",
            "description": "This value should be entered either as an ICD-9 code or as free text. If multiple diagnoses are recorded, they shall be separated by repeat delimiters (\\).",
            "nullable": true
          },
          "activeMeds": {
            "type": "string",
            "description": "This field is used for patient active medications or those suspected, in overdose situations. The generic name shall be used. This field is of use in interpretation of clinical results.",
            "nullable": true
          },
          "diet": {
            "type": "string",
            "description": "This optional field in free text should be used to indicate such conditions that affect results of testing, such as 16-hour fast (for triglycerides) and no red meat (for hemoccult testing).",
            "nullable": true
          },
          "pF1": {
            "type": "string",
            "description": "This is a text field for use by the practice; the optional transmitted text will be returned with the results.",
            "nullable": true
          },
          "pF2": {
            "type": "string",
            "description": "This is a text field for use by the practice; the optional transmitted text will be returned with the results.",
            "nullable": true
          },
          "admDates": {
            "type": "string",
            "description": "Admission and discharge dates, in YYYYMMDD format. The discharge date, when included, follows the admission date and is separated from it by a repeat delimiter (\\).",
            "nullable": true
          },
          "admStatus": {
            "type": "string",
            "description": "Admission status, represented by the following minimal list or by extensions agreed upon between the sender and receiver: OP (outpatient), PA (preadmit), IP (inpatient), ER (emergency room).",
            "nullable": true
          },
          "location": {
            "type": "string",
            "description": "The general clinic location or nursing unit, or ward or bed (or both) of the patient.",
            "nullable": true
          },
          "altCodeNature": {
            "type": "string",
            "description": "Nature of Alternative Diagnostic Code and Classifiers. It identifies the class of code or classifiers that are transmitted (e.g., DRGs, or in the future, AVGs [ambulatory visitation groups]).",
            "nullable": true
          },
          "altCode": {
            "type": "string",
            "description": "Alternative diagnostic codes and classifications (e.g., DRG codes) can be included in this field. The nature of the diagnostic code is identified in the AltCodeNature field. If multiple codes are included, they should be separated by repeat delimiters (\\). Individual codes can be followed by optional test descriptors (when the latter are present) and must be separated by component delimiters (^).",
            "nullable": true
          },
          "religion": {
            "type": "string",
            "description": "When needed, this value shall include the patient’s religion. Codes or names may be sent as agreed upon between the sender and the receiver. Full names of religions may also be sent as required.",
            "nullable": true
          },
          "maritalStatus": {
            "type": "string",
            "description": "When required, this value shall indicate the marital status of the patient as follows:<br>M = married<br>S = single<br>D = divorced<br>W = widowed<br>A = separated",
            "nullable": true
          },
          "isolationStatus": {
            "type": "string",
            "description": "Isolation codes indicate precautions that must be applied to protect the patient or staff against infection. Multiple precautions can be listed when separated by repeat delimiters (\\). Full text precautions may also be sent.",
            "nullable": true
          },
          "language": {
            "type": "string",
            "description": "The patient’s primary language. This may be needed when the patient is not fluent in the local language.",
            "nullable": true
          },
          "hospService": {
            "type": "string",
            "description": "The hospital service currently assigned to the patient. Both code and text may be sent when separated by a component delimiter (^).",
            "nullable": true
          },
          "hospInstitution": {
            "type": "string",
            "description": "The hospital institution currently assigned to the patient. Both code and text may be sent when separated by a component delimiter (^).",
            "nullable": true
          },
          "dosageCategory": {
            "type": "string",
            "description": "The patient dosage group. For example, A–ADULT, P1–PEDIATRIC (one to six months), P2–PEDIATRIC (six months to three years), etc.",
            "nullable": true
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "This record type contains information about an individual patient."
      },
...
    }
  }
}